EventRulesController: Move `addAction` to `EventRuleController`
